SVN_MERGE
Merged revisions 1038391,1038506,1038517,1038611,1038618,1038670,1038704,1038806,1038854,1038873,1038914 via svnmerge from 
https://vkrause@svn.kde.org/home/kde/trunk/KDE/kdepim

........
  r1038391 | tokoe | 2009-10-21 10:59:02 +0200 (Wed, 21 Oct 2009) | 7 lines
  
  Remove Scalix support from KDE PIM
  
  As long as we have no server for testing available
  having this code around makes no sense.
  The next step would be to create a ScalixProxyResource
  based on the new IMAP resource for Akonadi anyway.
........
  r1038506 | scripty | 2009-10-21 12:08:41 +0200 (Wed, 21 Oct 2009) | 1 line
  
  SVN_SILENT made messages (.desktop file)
........
  r1038517 | pokrzywka | 2009-10-21 12:16:21 +0200 (Wed, 21 Oct 2009) | 1 line
  
  add missing library to the link
........
  r1038611 | winterz | 2009-10-21 15:56:00 +0200 (Wed, 21 Oct 2009) | 3 lines
  
  compile warnings--
  MERGE: none
........
  r1038618 | pokrzywka | 2009-10-21 16:08:03 +0200 (Wed, 21 Oct 2009) | 1 line
  
  grrr, #warning is gcc specific... There are better ways to set tracepoints or comments in the code :-o
........
  r1038670 | winterz | 2009-10-21 17:42:18 +0200 (Wed, 21 Oct 2009) | 3 lines
  
  compile warning--
  MERGE: none
........
  r1038704 | djarvie | 2009-10-21 19:10:15 +0200 (Wed, 21 Oct 2009) | 2 lines
  
  Remove #warning
........
  r1038806 | mkoller | 2009-10-21 23:42:31 +0200 (Wed, 21 Oct 2009) | 2 lines
  
  make sure to change the window title not only when selecting IMAP folders
........
  r1038854 | jpike | 2009-10-22 06:33:21 +0200 (Thu, 22 Oct 2009) | 2 lines
  
  cryptopageplugin.cpp includes kleopatra headers which require the gpgme include path to be set.
........
  r1038873 | vkrause | 2009-10-22 09:09:58 +0200 (Thu, 22 Oct 2009) | 2 lines
  
  fix clean build
........
  r1038914 | scripty | 2009-10-22 11:37:39 +0200 (Thu, 22 Oct 2009) | 1 line
  
  SVN_SILENT made messages (.desktop file)
........

svn path=/branches/work/akonadi-ports/kdepim/; revision=1039004

+++ koeventview.cpp
@@ -252,12 +252,13 @@ bool KOEventView::processKeyEvent( QKeyEvent *ke )
     case Qt::Key_Alt:
       break;
     default:
-      mTypeAheadEvents.append( new QKeyEvent( ke->type(),
-                                              ke->key(),
-                                              ke->modifiers(),
-                                              ke->text(),
-                                              ke->isAutoRepeat(),
-                                              ke->count() ) );
+      mTypeAheadEvents.append(
+        new QKeyEvent( ke->type(),
+                       ke->key(),
+                       ke->modifiers(),
+                       ke->text(),
+                       ke->isAutoRepeat(),
+                       static_cast<ushort>( ke->count() ) ) );
       if ( !mTypeAhead ) {
         mTypeAhead = true;
         emit newEventSignal();
